Added checkout(...) and init(...) method
Checkout does exactly what it is called for: Git checkout a specific branch in a specific repository, also update Git submodules and Git LFS if necessary.
Init does a bunch of things: Based on the multibranch pipeline environment variables it checks if the current build is a feature branch or pull request build and therefore tries to create a merge build.