Added more methods for BitBucket integration
Method "projectName" to get the BitBucket project name, also added jUnit test.
Method "checkForOpenPullRequest" to check if a specific branch also exists as a pull request. This is a specific method for Jenkins integration, where feature branch builds can be "skipped" when a pull request exists for this specific feature branch build.